requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.7
django-bootstrap4==3.0.0
django-crispy-forms==1.11.2
django-filter==2.4.0
factory-boy==3.2.0
Faker==8.0.0
Pillow==8.1.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3     
